// src/shared/api/types/showCredentials.tsx
export interface ShowCredentialsResponse {
  cast: ShowCast[];
  crew: ShowCrew[];
  id: number;
}

export interface ShowCast {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path: string;
  character: string;
  credit_id: string;
  order: number;
}

export interface ShowCrew {
  adult: boolean;
  gender: number;
  id: number;
  known_for_department: string;
  name: string;
  original_name: string;
  popularity: number;
  profile_path?: string;
  credit_id: string;
  department: string;
  job: string;
}